Low: slapd: new ocf resource agent
I created a resource agent for slapd, and want to submit it to the project.
I know I should upload it as a patch, but I was having some difficulties
with that, so I'm doing it this way instead. If I should use the official
method or use different steps, please let me know.

I will happily modify the script or provide additional information to get
it included.